Update to Qt 6.9.3, fix compilation issues, and update CI/CD workflow

+++ src/hlam/ui/options/gameconfigurations/AutodetectScanResultsPage.cpp
@@ -137,7 +137,7 @@ void AutodetectScanResultsPage::OnStartScan()
 	const auto directories = gameDirectory.entryInfoList(QDir::Dirs, QDir::SortFlag::Name);
 
 	{
-		QProgressDialog dialog{"Scanning for games...", "Abort scan", 0, directories.size(), this};
+		QProgressDialog dialog{"Scanning for games...", "Abort scan", 0, static_cast<int>(directories.size()), this};
 		dialog.setWindowModality(Qt::WindowModal);
 		dialog.setMinimumDuration(2000);
